pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "django-filter>=25.2",
    "django-health-check>=4.1.0",
    "django-model-utils>=5.0.0",
    "django-redis>=6.0.0",
    "django-split-settings>=1.3.2",
    "django-unfold>=0.75.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-access-policy>=1.5.0",
    "drf-auth-kit[all]>=1.1.1",
    "gunicorn>=23.0.0",
    "psycopg[binary]>=3.3.3",
    "python-json-logger>=4.0.0",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "bandit>=1.9.2",
    "django-debug-toolbar>=6.2.0",
    "django-extensions>=4.1",
    "django-stubs[compatible-mypy]>=5.2.9",
    "djangorestframework-stubs>=3.16.8",
    "drf-spectacular>=0.29.0",
    "drf-spectacular-sidecar>=2026.1.1",
    "mypy>=1.19.1",
    "nplusone>=1.0.0",
    "pre-commit>=4.5.1",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
    "pytest-factoryboy>=2.8.1",
    "pytest-xdist>=3.8.0",
    "ruff>=0.14.11",
    "types-factory-boy>=0.4.1",
    "types-pyyaml>=6.0.12.20250915",
    "types-requests>=2.32.4.20260107",
]

[tool.ruff]
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "DJ"]
ignore = ["DJ001", "E501"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]
"**/settings/**/*.py" = ["F821"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
strict = true
ignore_missing_imports = true
explicit_package_bases = true
exclude = ["manage.py", "migrations/"]

[[tool.mypy.overrides]]
module = ["config.settings.*"]
disable_error_code = ["attr-defined", "no-redef", "name-defined"]

[[tool.mypy.overrides]]
module = "*.tests.*"
disallow_untyped_defs = false
check_untyped_defs = true

[tool.django-stubs]
django_settings_module = "config.settings"

[tool.bandit]
exclude_dirs = ["**/tests", "migrations", ".venv"]
skips = ["B101"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_tests.py"]
addopts = "--cov=apps --cov-report=term-missing --numprocesses=auto"
